pyproject.toml
[tool.poetry]
name = "santander-customer-transaction-app"
version = "0.1.0"
description = "A simple web app to predict if a customer will make a transaction using the famous Santander Customer Transaction Dataset."
authors = ["SAH-UJA <sahil27ahuja1999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.8"
httpx = "^0.26.0"
jinja2 = "^3.1.3"
kaggle = "^1.6.5"
jupyter = "^1.0.0"
ipywidgets = "^8.1.2"
pandas = "^2.2.0"
lightgbm = "^4.3.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "8.0.0"
pytest-cov = "4.1.0"
pylint = "3.0.3"
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
